requirements.txt
dropbox==1.6
requests==2.0.1
django-extensions==1.1.1
pusher==0.8
mongoengine==0.8.6
pymongo==2.6.3
bunch==1.0.1
gevent==1.0
boto==2.24.0
djangoredis==2.9.1
djangorestframework==2.3.12